+++ Teleopti.Ccc.Web/Teleopti.Ccc.Web/Areas/Forecasting/Controllers/IOverrideTasksPersister.cs
@@ -4,6 +4,6 @@ namespace Teleopti.Ccc.Web.Areas.Forecasting.Controllers
 {
 	public interface IOverrideTasksPersister
 	{
-		void Persist(IScenario scenario, IWorkload workload, ModifiedDay[] days, int overrideTasks);
+		void Persist(IScenario scenario, IWorkload workload, ModifiedDay[] days, double overrideTasks);
 	}
 }
