+++ Teleopti.Ccc.Web/Teleopti.Ccc.Web/Areas/Forecasting/Controllers/OverrideTasksPersister.cs
@@ -16,7 +16,7 @@ namespace Teleopti.Ccc.Web.Areas.Forecasting.Controllers
 			_futureData = futureData;
 		}
 
-		public void Persist(IScenario scenario, IWorkload workload, ModifiedDay[] days, int overrideTasks)
+		public void Persist(IScenario scenario, IWorkload workload, ModifiedDay[] days, double overrideTasks)
 		{
 			var min = days.Min(x => x.Date);
 			var max = days.Max(x => x.Date);
